refactor static utils to a separate file
